requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
absl-py=2.1.0=pypi_0
altair=5.2.0=pypi_0
annotated-types=0.6.0=pypi_0
anyio=4.2.0=pypi_0
aom=3.8.1=h63175ca_0
asgiref=3.7.2=pypi_0
assemblyai=0.20.2=pypi_0
asttokens=2.4.1=pyhd8ed1ab_0
astunparse=1.6.3=pypi_0
attrs=23.2.0=pypi_0
audioread=3.0.1=pypi_0
backcall=0.2.0=pyh9f0ad1d_0
blas=1.0=mkl
blis=0.7.11=pypi_0
bzip2=1.0.8=hcfcfb64_5
ca-certificates=2024.2.2=h56e8100_0
cachetools=5.3.3=pypi_0
cairo=1.18.0=h1fef639_0
catalogue=2.0.10=pypi_0
certifi=2024.2.2=pyhd8ed1ab_0
cffi=1.16.0=pypi_0
chardet=5.2.0=pypi_0
charset-normalizer=3.3.2=pyhd8ed1ab_0
click=8.1.7=pypi_0
cloudpathlib=0.16.0=pypi_0
colorama=0.4.6=pyhd8ed1ab_0
colorlog=6.8.2=pypi_0
comm=0.2.1=pyhd8ed1ab_0
confection=0.1.4=pypi_0
contourpy=1.2.0=pypi_0
cuda-cccl=12.4.99=0
cuda-cudart=11.8.89=0
cuda-cudart-dev=11.8.89=0
cuda-cupti=11.8.87=0
cuda-libraries=11.8.0=0
cuda-libraries-dev=11.8.0=0
cuda-nvrtc=11.8.89=0
cuda-nvrtc-dev=11.8.89=0
cuda-nvtx=11.8.86=0
cuda-profiler-api=12.4.99=0
cuda-runtime=11.8.0=0
cudatoolkit=11.2.2=h7d7167e_13
cudnn=8.1.0.77=h3e0f4f4_0
cycler=0.12.1=pypi_0
cymem=2.0.8=pypi_0
datapane=0.17.0=pypi_0
dav1d=1.2.1=hcfcfb64_0
debugpy=1.8.0=py39h99910a6_1
decorator=4.4.2=pypi_0
distro=1.9.0=pypi_0
django=4.2.10=pypi_0
django-appconf=1.0.6=pypi_0
django-compressor=4.4=pypi_0
django-libsass=0.9=pypi_0
django-sass-processor=1.4=pypi_0
djangorestframework=3.15.0=pypi_0
dominate=2.9.1=pypi_0
en-core-web-sm=3.7.1=pypi_0
errant=3.0.0=pypi_0
executing=2.0.1=pyhd8ed1ab_0
expat=2.5.0=h63175ca_1
ffmpeg=6.1.1=gpl_hb766fab_104
filelock=3.13.1=py39haa95532_0
flatbuffers=24.3.7=pypi_0
font-ttf-dejavu-sans-mono=2.37=hab24e00_0
font-ttf-inconsolata=3.000=h77eed37_0
font-ttf-source-code-pro=2.038=h77eed37_0
font-ttf-ubuntu=0.83=h77eed37_1
fontconfig=2.14.2=hbde0cde_0
fonts-conda-ecosystem=1=0
fonts-conda-forge=1=0
fonttools=4.50.0=pypi_0
freetype=2.12.1=hdaf720e_2
fsspec=2024.3.0=pypi_0
fuzzywuzzy=0.18.0=pypi_0
gast=0.4.0=pypi_0
gettext=0.21.1=h5728263_0
gmpy2=2.1.2=py39h7f96b67_0
google-auth=2.28.2=pypi_0
google-auth-oauthlib=0.4.6=pypi_0
google-pasta=0.2.0=pypi_0
gramformer=1.0=pypi_0
graphite2=1.3.13=1000
grpcio=1.62.1=pypi_0
h11=0.14.0=pypi_0
h5py=3.10.0=pypi_0
harfbuzz=8.3.0=h7ab893a_0
httpcore=1.0.2=pypi_0
httpx=0.26.0=pypi_0
huggingface-hub=0.21.4=pypi_0
icu=73.2=h63175ca_0
idna=3.6=pypi_0
imageio=2.34.0=pypi_0
imageio-ffmpeg=0.4.9=pypi_0
importlib-metadata=7.0.1=pyha770c72_0
importlib-resources=6.3.1=pypi_0
importlib_metadata=7.0.1=hd8ed1ab_0
intel-openmp=2023.1.0=h59b6b97_46320
ipykernel=6.29.1=pyha63f2e9_0
ipynbname=2023.2.0.0=pypi_0
ipython=8.12.0=pyh08f2357_0
jedi=0.19.1=pyhd8ed1ab_0
jinja2=3.1.3=py39haa95532_0
joblib=1.3.2=pypi_0
jpeg=9e=h2bbff1b_1
jsonschema=4.21.1=pypi_0
jsonschema-specifications=2023.12.1=pypi_0
jupyter_client=8.6.0=pyhd8ed1ab_0
jupyter_core=5.7.1=py39hcbf5309_0
keras=2.10.0=pypi_0
keras-preprocessing=1.1.2=pypi_0
kiwisolver=1.4.5=pypi_0
langcodes=3.3.0=pypi_0
lazy-loader=0.3=pypi_0
lerc=3.0=hd77b12b_0
levenshtein=0.25.0=pypi_0
libclang=18.1.1=pypi_0
libcublas=11.11.3.6=0
libcublas-dev=11.11.3.6=0
libcufft=10.9.0.58=0
libcufft-dev=10.9.0.58=0
libcurand=10.3.5.119=0
libcurand-dev=10.3.5.119=0
libcusolver=11.4.1.48=0
libcusolver-dev=11.4.1.48=0
libcusparse=11.7.5.86=0
libcusparse-dev=11.7.5.86=0
libdeflate=1.17=h2bbff1b_1
libexpat=2.5.0=h63175ca_1
libffi=3.4.2=h8ffe710_5
libglib=2.78.3=h16e383f_0
libiconv=1.17=hcfcfb64_2
libjpeg-turbo=2.0.0=h196d8e1_0
libnpp=11.8.0.86=0
libnpp-dev=11.8.0.86=0
libnvjpeg=11.9.0.86=0
libnvjpeg-dev=11.9.0.86=0
libopus=1.3.1=h8ffe710_1
libpng=1.6.42=h19919ed_0
librosa=0.10.1=pypi_0
libsass=0.23.0=pypi_0
libsodium=1.0.18=h8d14728_1
libsqlite=3.44.2=hcfcfb64_0
libtiff=4.5.0=h8a3f274_0
libuv=1.44.2=h2bbff1b_0
libwebp-base=1.3.2=h2bbff1b_0
libxml2=2.12.5=hc3477c8_0
libzlib=1.2.13=hcfcfb64_5
llvmlite=0.42.0=pypi_0
lxml=4.9.4=pypi_0
lz4-c=1.9.4=h2bbff1b_0
markdown=3.6=pypi_0
markupsafe=2.1.5=pypi_0
matplotlib=3.8.3=pypi_0
matplotlib-inline=0.1.6=pyhd8ed1ab_0
micawber=0.5.5=pypi_0
mkl=2023.1.0=h6b88ed4_46358
mkl-service=2.4.0=py39h2bbff1b_1
mkl_fft=1.3.8=py39h2bbff1b_0
mkl_random=1.2.4=py39h59b6b97_0
moviepy=1.0.3=pypi_0
mpc=1.1.0=h7edee0f_1
mpfr=4.0.2=h62dcd97_1
mpir=3.0.0=hec2e145_1
mpmath=1.3.0=py39haa95532_0
msgpack=1.0.7=pypi_0
multimethod=1.11.2=pypi_0
murmurhash=1.0.10=pypi_0
nest-asyncio=1.6.0=pyhd8ed1ab_0
networkx=3.1=py39haa95532_0
numba=0.59.0=pypi_0
numpy=1.26.4=py39h055cbcc_0
numpy-base=1.26.4=py39h65a83cf_0
oauthlib=3.2.2=pypi_0
openai=1.11.1=pypi_0
opencv-python=4.9.0.80=pypi_0
openh264=2.4.1=h63175ca_0
openjpeg=2.4.0=h4fc8c34_0
openssl=3.2.1=hcfcfb64_0
opt-einsum=3.3.0=pypi_0
packaging=23.2=pyhd8ed1ab_0
pandas=1.5.3=pypi_0
parso=0.8.3=pyhd8ed1ab_0
pcre2=10.42=h17e33f8_0
pickleshare=0.7.5=py_1003
pillow=10.2.0=py39h2bbff1b_0
pip=24.0=pyhd8ed1ab_0
pixman=0.43.2=h63175ca_0
platformdirs=4.2.0=pyhd8ed1ab_0
pooch=1.8.0=pypi_0
preshed=3.0.9=pypi_0
proglog=0.1.10=pypi_0
prompt-toolkit=3.0.42=pyha770c72_0
prompt_toolkit=3.0.42=hd8ed1ab_0
protobuf=3.19.6=pypi_0
psutil=5.9.8=py39ha55989b_0
pure_eval=0.2.2=pyhd8ed1ab_0
pyarrow=15.0.1=pypi_0
pyasn1=0.5.1=pypi_0
pyasn1-modules=0.3.0=pypi_0
pycparser=2.21=pypi_0
pydantic=2.6.1=pypi_0
pydantic-core=2.16.2=pypi_0
pydub=0.25.1=pypi_0
pygments=2.17.2=pyhd8ed1ab_0
pynacl=1.5.0=pypi_0
pyparsing=3.1.2=pypi_0
python=3.9.18=h4de0772_1_cpython
python-dateutil=2.8.2=pyhd8ed1ab_0
python-levenshtein=0.25.0=pypi_0
python_abi=3.9=4_cp39
pytorch=2.2.1=py3.9_cuda11.8_cudnn8_0
pytorch-cuda=11.8=h24eeafa_5
pytorch-mutex=1.0=cuda
pytz=2024.1=pypi_0
pywin32=306=py39h99910a6_2
pyyaml=6.0.1=py39h2bbff1b_0
pyzmq=25.1.2=py39h7eaf5a6_0
rapidfuzz=3.6.2=pypi_0
rcssmin=1.1.1=pypi_0
referencing=0.34.0=pypi_0
regex=2023.12.25=pypi_0
requests=2.31.0=py39haa95532_1
requests-oauthlib=1.4.0=pypi_0
rjsmin=1.2.1=pypi_0
rpds-py=0.18.0=pypi_0
rsa=4.9=pypi_0
safetensors=0.4.2=pypi_0
scikit-learn=1.4.0=pypi_0
scipy=1.12.0=pypi_0
sentencepiece=0.2.0=pypi_0
setuptools=69.0.3=pyhd8ed1ab_0
six=1.16.0=pyh6c4a22f_0
smart-open=6.4.0=pypi_0
sniffio=1.3.0=pypi_0
soundfile=0.12.1=pypi_0
soxr=0.3.7=pypi_0
spacy=3.7.4=pypi_0
spacy-legacy=3.0.12=pypi_0
spacy-loggers=1.0.5=pypi_0
sqlite=3.44.2=hcfcfb64_0
sqlparse=0.4.4=pypi_0
srsly=2.4.8=pypi_0
stack_data=0.6.2=pyhd8ed1ab_0
svt-av1=1.8.0=h63175ca_0
sympy=1.12=py39haa95532_0
tbb=2021.8.0=h59b6b97_0
tensorboard=2.10.1=pypi_0
tensorboard-data-server=0.6.1=pypi_0
tensorboard-plugin-wit=1.8.1=pypi_0
tensorflow=2.10.1=pypi_0
tensorflow-estimator=2.10.0=pypi_0
tensorflow-io-gcs-filesystem=0.31.0=pypi_0
termcolor=2.4.0=pypi_0
thinc=8.2.3=pypi_0
threadpoolctl=3.2.0=pypi_0
tk=8.6.13=h5226925_1
tokenizers=0.15.2=pypi_0
toolz=0.12.1=pypi_0
torchaudio=2.2.1=pypi_0
torchvision=0.17.1=pypi_0
tornado=6.3.3=py39ha55989b_1
tqdm=4.66.1=pypi_0
traitlets=5.14.1=pyhd8ed1ab_0
transformers=4.38.2=pypi_0
typer=0.9.0=pypi_0
typing_extensions=4.9.0=pyha770c72_0
tzdata=2023.4=pypi_0
ucrt=10.0.22621.0=h57928b3_0
urllib3=2.2.0=pypi_0
vc=14.3=hcf57466_18
vc14_runtime=14.38.33130=h82b7239_18
vega-datasets=0.9.0=pypi_0
vs2015_runtime=14.38.33130=hcb4865c_18
wasabi=1.1.2=pypi_0
wcwidth=0.2.13=pyhd8ed1ab_0
weasel=0.3.4=pypi_0
websockets=12.0=pypi_0
werkzeug=3.0.1=pypi_0
wheel=0.42.0=pyhd8ed1ab_0
wrapt=1.16.0=pypi_0
x264=1!164.3095=h8ffe710_2
x265=3.5=h2d74725_3
xz=5.2.6=h8d14728_0
yaml=0.2.5=he774522_0
zeromq=4.3.5=h63175ca_0
zipp=3.17.0=pyhd8ed1ab_0
zlib=1.2.13=hcfcfb64_5
zstd=1.5.2=h19a0ad4_0
